requirements.txt
Django==5.1.3
requests==2.32.3
PyGithub==2.5.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.8.2
Brotli==1.1.0
dj-database-url==2.3.0
social-auth-app-django==5.4.2
Jinja2==3.1.4
celery[redis]==5.4.0
django-celery-results==2.5.1
sendgrid==6.11.0
humanize==4.11.0
djangoql==0.18.1
requests-cache==1.2.1
Markdown==3.7
PyYAML==6.0.2
tqdm==4.67.1
